Allow for filtering of Blade engine function
The actual function or method used to render blade templates can vary;
this allows the user to filter the callable that acf-gutenblocks uses
to verify it *can* render, and to do the actual rendering. If the filter
is not used, then it behaves exactly as it did before (i.e. it uses
\App\template()).